bundling resources in single js and css files - decreasing network request for js and css from 79 to 2 - reducing files size with minification and maximum compression
- JS -
2,5M	bundle.js
1,4M	bundle.min.js
408K	bundle.min.js.gz

- CSS -
212K	bundle.css
200K	bundle.min.css
32K	bundle.min.css.gz

worth it, isn't? :D

To activate the use of pre-compressed files in nginx, simply add
```
  gzip_static on;
```
in the appropriate location block.

To run the build:
$ npm run build

This is a quick & dirty build script, in the long run
you might want to consider a specialized builder tool such as brunch (http://brunch.io) to get dependency management via npm and watch changes features