QuerySamples parser: remove meaningless repetitions
An "exactly 1" reprtition does not do anything. One match is always
done.

An "one or more" reprtition at the end of the pattern does not make
a difference. One match is always done. The regex can succeed then.
There is no need to look for "possibly more" when what we have is
already anough.

Change-Id: I2a68b0370a50ddec189eafd67ade55d8037cd45d