QuerySamples parser: replace [^] with more robust [\s\S]
Per argument given in http://stackoverflow.com/a/16119722

Quote: "I recommend against using [^] anyway. On one hand, JavaScript
is the only flavor I know that supports that idiom, and even there it's
used nowhere near as often as [\s\S]. On the other hand, most other
flavors let you escape the ] by listing it first. In other words, in
JavaScript [^][^] matches any two characters, but in .NET it matches
any one character other than ], [, or ^."

The {{Q|…|…}} template is special because it is not expected to contain
these characters, so we can safely build a meaningfull negative
character class that does not trigger backracking.

Change-Id: I27ce465e5389a8b70991ac3ac80a129099cbce5a