Fix node font color for image and dot nodes
Since for those nodes the text is printed below the node, not within the
colored area, their font color should always be black. We have to change
this in two locations, since we can't know if the rgb column will be
processed before the column that turns the node into a number/image node
or vice versa.

Bug reported by Jonas Kress in a comment on Chang 305189.

Change-Id: I0978859f35b64accc2007507adab4e5a1f2e9647