Refactor result error message parsing
This patch now tries to match errors as strings
before falling back to the failsafe regex which will
output some sort of shorter error for the progress bar.

This also adds a new error type 'malformed query'
to the list of recognized errors.

This allows the occourance of this error to be tracked.

Change-Id: I8edd8bdd159326ddca7ec2c65b623987bee20a2b