Replace problematic atob/btoa with encode/decodeURIComponent
This is a direct follow up to I58cd6e8. Note that this was not an issue
specific for Firefox. The URL encoding-decoding chain was not correct
in all browsers. It's just that Chrome does have othe rencoding magic in
place that tries other things when it encounters something unencoded in
a URL. Now there is enforced encoding via encodeURIComponent and the browser
does not need to do magic.

Bug: T149655
Change-Id: I882d99b844dd03b48508815a8bb8ea2e70726fdc