Fix T154194: bubble chart url off-by-one error
The bubble chart iterates over all rows and columns, but barely uses the
column iteration: the item is filled in almost completely as soon as the
next row is seen, accessing the correct indexes in the full row right
away. This works for label, number, rgb, but not URL – the URL is
actually searched in all columns.

Previously, the assumption was that the URL column is the first one: in
that case, the url variable is set as soon as the next row is seen, and
then the item is immediately created with the value of the url variable.
But if the URL column isn’t the first one, this means that the item will
instead be populated with the previous row’s URL, which is still
leftover in the url variable. The url variable is then set to the
current row’s URL when that column is reached, but by that point the
item isn’t updated, so the url variable is only read by the next item
again.

Instead, get the last item and set its URL after the rollover. If the
URL column is the first one, then the rollover and the URL setting will
happen in the same iteration, but if it isn’t, the URL is still set on
the correct item.

Change-Id: I2fe1b4b7a7fc26468f2fdeea1e54ed6f850123fb