Add a few more SPARQL predicates
wikibase:Property is the rdf:type of property entities, which are also
the subject of many of the property predicates listed just above the new
entity types section (wikibase:claim etc.).

wikibase:Item and wikibase:Statement are the rdf:type of items and
statements, respectively, but currently disabled on WDQS for performance
reasons.

wikibase:WikibaseProperty and wikibase:Math are two possible values for
wikibase:propertyType, the type of a property, which were previously
missing.

Change-Id: Ia478ad97b9dd3a278869c2062d4b62a85afc3a74