Expand timeline ranges on hover
If a query result contains two points in time (start and end), we
display them as a range in the timeline. Unfortunately, this means that
the content of the range (additional result columns) can be partially or
even completely hidden if the range is too short.

To remediate this, we unset the width of the range element when it is
hovered: this lets the browser adjust the width so that the content fits
inside it. (The range is absolutely positioned with a “left:” property,
so on resize the right edge of the range moves while the left edge stays
in place, creating a sort of “flyout” effect.)

However, this only needs to be done for ranges that are actually too
short to fit their content: we don’t want this to happen to ones that
are wide enough. But we cannot test this via CSS, so unsetting the width
also lets the browser shrink long ranges that are “too big” for their
content.

To avoid this, we also set the min-width of each range to its width
assigned by the timeline: ranges that are too short are unaffected, but
ranges that are too long are not shrunk on hover.

Bug: T163984
Change-Id: I965cf7043c21e2f2dbe26ed9b5547d06732547f6