Fixed Table-Width Error in QueryExampleDialog
The Width of the first column in the the QED is determined by its
longest entry. When filtering these entry, or this list contains
only short entries, the right columns (containing the glyphicons
for source and preview) shift to the left, i.e. their colums
"grow". By explicitly setting a column width, this bug is fixed.

Change-Id: I06ff1ffac2ccb66f0e04e9b9dfd4f492c4e3bc85