Only send queries via POST if GET fails
Queries sent via POST are not cached, so we should only fall back to
POST if GET does not work due to the query length. We detect that case
by checking if the response headers are completely empty, which
according to MDN [1] happens in case of network error, but should never
happen otherwise.

For more details about GET vs. POST and query length limitations, see
commit f02b6c2b0d / change Ic770e25bb4.

[1]: https://developer.mozilla.org/en-US/docs/Web/API/XMLHttpRequest/getAllResponseHeaders

Change-Id: Ic3ea53c50edf2be30dbcce86a67923b621887c5b