Add rel="noopener" to many target="_blank" links
Without rel="noopener", the target page has access to the original
window via `window.opener` and can access and modify some properties;
for example, try running this query and clicking the result link:

    SELECT (<https://lucaswerkmeister.de/no-noopener.html> AS ?x) WHERE {}

This commit adds rel="noopener" to many, but not all links with
target="_blank". In general, the attack surface on query.wikidata.org is
small, since there is little to be gained from impersonating the website
(it does not have, for instance, any login information that the user
might enter). Internal links don’t always need rel="noopener", since the
target can be trusted; on the other hand, this commit does add
rel="noopener" to some of them, just for consistency in the HTML with
external links that should have rel="noopener".

Change-Id: I4d5c5d6949222ea2c971395d3155e87f0d1acf36