Store variable name in data attribute instead of class
Using the class can result in unintended styling on the variable span;
for instance, the generic name ?item results in a framed box, since item
is the CSS class for Masonry items (Image result browser).

We use .attr() instead of jQuery’s .data() since .data() does not store
its data in data attributes and is therefore not queryable via .find().

Using data attributes instead of classes also makes it unnecessary to
strip the question mark from the variable name, as long as the attribute
selector quotes the value.

Change-Id: I4376d17227dbc95ff8132ac505f386b7b538bb89