Move comment handling from VisualEditor to SparqlQuery
SparqlQuery already had a _queryComments member, but never used it –
VisualEditor had its own _queryComments and managed them independently
of the SparqlQuery. This commit moves the comments handling to
SparqlQuery: they are extracted in parse() and added back in
getQueryString(). SparqlQuery also gains a function to extract the
content of a “marker” comment, like the #TEMPLATE= comments.

VisualEditor._cleanQueryPrefixes() needs a slight adjustment because its
query argument now already includes the comments, before the prefixes;
as a consequence, after replacing the prefix lines with empty strings,
the resulting blank lines are no longer at the beginning of the string
and therefore no longer removed by trim().

Tests are added to check that getQueryString() includes the comments,
and that getCommentContent() works.

Change-Id: I624172e1b44033766fa41d165f8f713187c7194e