Add QueryTemplate.parse and .getHtml functions
parse() takes a query and attempts to extract the template. getHtml()
renders the template.

Tests are added to check that getHtml() returns a correctly rendered
template, including a test for using the same variable multiple times in
a template (which was not previously supported).

This was originally part of Ic362ea1087.

Bug: T167868
Change-Id: I91f31272870a574721667e301f103eb5e78de01f