Fix variable is removed even if it still is used
If multiple filters restricted the same variable (e. g. instance of X
with part of Y), removing one of the filters would remove the variable
from the SELECT clause. This should only happen when the last filter for
a variable is removed, when the variable is actually no longer bound
within the query.

Change-Id: I899b176b289e6b2afa338a15fe879eb935a997dd