Support setting edge labels in graph view
The special variable name ‘?edgeLabel’ is used for the edge label. To
support this, the edge variable is persisted across iterations over the
results, just like the node variable. This is not without problems: if
the label variable appears before the variable for the node to link to,
then edge refers to the previous edge, and edge labels will be shifted
(on the wrong edge). I think this is acceptable for now, since unusual
variable orders can already cause other problems (e. g., the image
variable has to appear after the first node variable).

Bug: T168715
Change-Id: Iebc3288a34d523de61b332a0ab7e4a1f752ad1f1