Make wkt literal matching more lenient
I haven’t been able to find a standard document on this (GeoSPARQL [1]
only specifies that a geo:wktLiteral contains a WKT String according to
ISO 191215-1, which is not freely available), but the
http://linkedgeodata.org/sparql service (which is part of the federation
whitelist) spells these literals all uppercase, i.e. POINT(1 2), so it
seems that the Point should be matched case-insensitively.

The same endpoint also contains other kinds of literals, e. g.
LINESTRING, so we should also not assume that the regular expression
will always match, and instead handle the non-matching case more
gracefully.

[1]: http://www.opengeospatial.org/standards/geosparql

Change-Id: I7ceb1b94a68228ba94733ec86017218966b4f152