Use wellknown to parse geo:wktLiteral values
_extractLongLat is changed to _extractGeoJson and uses wellknown to
parse the WKT literal instead of extracting point coordinates via
regular expression. This gives us support for other kinds of values, for
instance linestrings or polygons; such values are not present on
Wikidata, but may be obtained via a federated SPARQL query against a
GeoSPARQL endpoint, e. g. LinkedGeoData.org.

_createMarkerGroups is updated to store the returned GeoJSON in a
GeoJSON layer, using the pointToLayer option to change the Leaflet
default pin to a simple point marker. The popup is bound to any kind of
feature, including non-point features.

The tests for _extractLongLat are updated to account for the rename and
the changed return type, and a test for a simple non-point geometry is
added.

Change-Id: I9c069eff00ac816fa345aa1b8ab874032730a725