Parse geo:wktLiteral according to GeoSPARQL
Instead of checking whether Point(…) and the Wikidata Earth globe occur
anywhere within the value, split it according to GeoSPARQL [1], Req 10.
Additionally, use the default coordinate reference system according to
Req 11 and also accept it as a system similar to GLOBE_EARTH; for this
purpose, a new list EARTH_LONGLAT_SYSTEMS of such suitable systems is
added, which other systems can be added to later if needed. (The axis
order is significant, because according to Req 12, the interpretation of
the WKT literal depends on the axis order of the reference system.
Fortunately, Wikidata, CRS84, and GeoJSON all use longitude-latitude
order, so we do not need to worry about flipping axes yet.)

_extractGlobe is removed, since its job is essentially done by
_splitWktLiteral now; _checkCoordinate simply uses the full
_extractLongLat function to check if a value is a coordinate or not.

Tests for _splitWktLiteral and _extractLongLat are added.

[1]: http://www.opengeospatial.org/standards/geosparql

Change-Id: I8310b1d69b0f01a89c9c20bcfb8ec0996298e791