Show query template by default on embed.html
If a query template is set, the Query Helper is useful enough that we
can show it immediately, without waiting for the user to click the edit
button. (The button can still be used to toggle the Helper, this just
changes the default state.)

The Query Helper is toggled after a short timeout to prevent showing the
template with unrendered variables to the user if possible.

Change-Id: Iaebe5aa7107746b0c9a9c04b36f9f2551733020f