Load translated examples if available
This attempts to load a translated subpage of the examples for the
current language before falling back to the main page (in English).

Note that there is no fallback logic for languages – any language that
does not have a translated subpage immediately falls back to English.

Changing a language also does not yet update the loaded examples, so you
generally have to reload the page after changing the language to get
translated examples.

Bug: T169607
Change-Id: Ic1d98e4793cd8e3b7a9205b559f12b9a5082adad