Make query helper and editor resizable
A splitter is added between query helper and query editor; dragging the
splitter resizes the query helper, giving it more or less space relative
to the query editor.

To achieve this, the main query form is changed to use flexbox instead
of a table. The jQuery plugin jquery-resizable is used to make the query
helper resizable (the npm package is called jquery-resizable-dom because
jquery-resizable was already taken). The CodeMirror editor does not
respond directly to the size of its container arranged by the Flex
layout, so it is explicitly resized when the drag operation terminates.

On small viewports, the container that contains the query helper and
editor switches to a vertical view (horizontal split).

Change-Id: Ia76bed3b4447729923abda2f12cc80f5e514e70c