Adapt query helper min-width to content
The tables inside the query helper can grow wider than the query helper
itself, which looks ugly when the query helper is resized past a certain
point. Dynamically set its min-width to the maximum width of the tables
(plus their offset from the left side of the query helper) to account
for this.

Note that when we adjust the min-width after the query has been edited,
all entity selectors in the query helper are without label, so the
minimum width we derive can be too small. However, this is still better
than nothing. (A label cache might mitigate this.)

(There is apparently a CSS solution for this – min-width: fit-content –
but it’s highly experimental and not yet available in any browser
without prefix.)

Change-Id: I27bd4cc5eddfc6b9fa97bb255db8ba792d33af53