Auto hide and show query box on scroll
This reduces the amount of scrolling needed when switching between the
editor and the query result, which is especially convenient when the
query result consumes scroll events (Map view).

The result navbar is moved out of the query box so that it is still
visible when the query box is hidden.

Change-Id: Iced825177110be330e40333b5e85ad7b3c00103c