Inline splitter background image as base64
Specifying a URL in node_modules/ probably doesn’t work in a deployment
build, and the file is small, so just include it in the CSS file
directly in base64.

The background directive is also split up so that comments can be added
for each part separately (indicating the image file name, and the
Wikimedia Style Guide color name).

Change-Id: I8a2c6a8faf881dc6de84b7a33452d1d995e3b1a3