Restrict query helper width
In CSS, set a maximum width of 80% of the viewport width. This should
ensure that the splitter is always still in view, so the user can still
resize the query helper even if it’s very wide.

In JS, cap the minimum width we set for the query helper at 50% of the
viewport width. Being blocked from making the query helper any smaller
is fairly annoying, so this cap needs to be way below the 80% from CSS.

Change-Id: I24aef51030478a1b67710367c1f1167f2ca6e240