Add support for SPARQL2 template in template data
With recent changes to the SPARQL2 template, for some reason it looks
just like the SPARQL template in the mw-data, so we can parse it with
the same code as the SPARQL template.

The old code for parsing the SPARQL2 template (using data.body.extsrc)
is removed, because without removing it some (but not all) SPARQL2 query
examples are loaded twice for some reason.

Bug: T172259
Change-Id: I061d376da3d163b34f89470bebced87490eec765