Show geoshapes in Map view
CoordinateResultBrowser._extractGeoJson is turned into a function that
delegates to either _extractGeoJsonWktLiteral (the old _extractGeoJson)
or _extractGeoJsonCommonsMap (new function), depending on the object
passed in. It can also return null, indicating that the value doesn’t
yield GeoJSON. Two new functions _isWktLiteral and _isCommonsMap are
used by both _extractGeoJson and _checkCoordinate to check if a value
can yield GeoJSON, and (for _extractGeoJson) how to extract it.

_extractGeoJson and its callees (_createMarkerGroups and draw) are also
turned into asynchronous functions, returning jQuery Promises, because
we shouldn’t fetch geoshapes synchronously. The tests are updated to
accommodate this change.

There are some open problems with this implementation:

* Map data is retrieved via the API, which is ugly. However, until
  T177966 is resolved, no other way is available. (This also means that
  responses are cached according to the maxage we specify in the API
  request, but never invalidated, so users may see slightly stale
  shapes.)
* The map is only drawn when all shapes have been loaded. It would be
  nicer to draw shapes as soon as they come in.

However, these seem acceptable for now, and can be fixed after the
Wikidata birthday.

Bug: T163910
Change-Id: I0188d31bb39c3e23d3bb65c17b492b1ca30ed063