Add support for per-result colors in map view
This adds support for an ?rgb column in the map view, just like in the
bubble chart and graph views. The ?rgb color takes precedence over the
?layer color; results without a valid ?rgb color still get the default
color of the layer, while the layers view will use the last seen ?rgb
color for a layer (or the default color if no ?rgb color was ever seen
for that layer), under the assumption that query authors will set the
same (or at least similar) ?rgb color(s) for all results in a layer.

(This feature is intended especially for geoshapes, but works
independently of geoshape support.)

Bug: T163910
Change-Id: Icf211d50ec05ae5ad139716b1b1e3c8ccf28de30