Moved CSS to LESS, added grunt support
The grunt build generates style.css from style.less. Additionally, for
local development, a section is added to index.html which compiles the
stylesheet in the client, so that after just `npm install` it is
possible to open index.html in the browser and have it working. In the
build, this section is discarded by htmlmin.

Change-Id: Ie4720ac9f212ced0d36ce54bc1a6c7ab294daada