Add support for message arguments to _i18n
We have a wrapper around _i18n in several places which allows us to use
messages even if $.i18n is not available (e. g. because some components
of the query service UI are reused elsewhere). This adds support for
message arguments to that wrapper, in a single array as the (optional)
third _i18n parameter.

(Note that there are two more copies of _i18n in queryHelper/, which are
not updated since they work slightly differently.)

Change-Id: I86324319f8fc75c667b2f996958ab8201e380a78