Show progress message while loading map data
As the promises for the map layers get resolved, track what ratio of the
promises is done overall. Concurrently, refresh a status message five
times per second, communicating that ratio to the user as percentage.

This is only relevant while loading geoshapes, since WKT literals are
parsed synchronously and returned immediately.

Change-Id: Ib30f7185ac2bed618511749ab6482bc450c2f220