Restore focus to editor after selecting example
Bootstrap restores focus to the originally opening button for a modal
after closing it, which is a bad idea for the query examples dialog
because a very common action after selecting a query example is to run
it with Ctrl+Enter, which also reopens the examples dialog. Instead,
focus the editor, which is also useful for immediately editing the
example query.

Bootstrap’s focus-restoring logic is wrapped in several layers of event
handlers, and I can’t figure out any way to register an event handler of
our own after Bootstrap’s focus-restoring event handler, so we use a
setTimeout instead to get our code to run after Bootstrap’s focus
restorer. It’s not ideal – sometimes you can see a brief flash of focus
on the examples button before we move focus to the editor – but I can’t
figure out any better solution.

Bug: T178089
Change-Id: I24613f225ee56133deceec6a8f3048ee62139c7b