Add CodeSamples tests
The tests consist of a query and the expected code in each language, all
placed in a folder under wikibase/tests/queryService/api/code-examples/
named after the test case. (This commit adds only a single test case,
“simple”; later commits can add more.)

A symbolic link is added so that the CodeSamples API can still load the
samples when running from wikibase/tests/ instead of the repository
root. Unfortunately this does not work in a browser, so the test is
disabled if it’s running in a browser. (Firefox’ same-origin policy for
the file:// protocol quite sensibly prohibits access to parent
directories of the document; Chrome/Chromium does not allow scripts to
access file:// URIs at all unless launched with
--allow-file-access-from-files.)

Change-Id: Ia751a6902c11abbe76d0f758ac838d37a152debf