Correctly float close button in RTL languages
In languages written right-to-left, the Query Helper’s “close” button
should be on the left side of the dialog, not on the right. This can be
done using the “inline-end” float value from CSS Logical Properties and
Values Level 1 [1]; unfortunately, this is currently only supported by
Firefox (since version 55), and other browsers will fall back to
Bootstrap’s float: right.

[1]: https://drafts.csswg.org/css-logical/#float-clear

Bug: https://phabricator.wikimedia.org/T178057
Change-Id: I23f8b6cb5256bcc0389ecd1a260ca127c578d8a5