Add option for layer column(s) of map view
The “layer” option can be either a single string or an array of strings,
specifying the selected variable(s) to use for the layer of a result
(the first nonempty variable in a result is used for the actual layer).
The variable must be specified with a leading question mark – other
values are reserved for future expansion.

Example:

    #defaultView:Map{"layer": "?partyLabel"}

Change-Id: Ided10bb8c8ae279ab51e245b1df95f85e6fc0bd1