Use full OpenStreetMap domain names
The osm.org domain was described as “not officially supported” in 2014 [1],
and while it is nowadays listed on the “Domain names” wiki page [2],
it’s still just a redirect, and we may as well get rid of it.

The main openstreetmap.org domain also redirects to www.openstreetmap.org,
so let’s use the www. subdomain right away.

[1]: https://help.openstreetmap.org/questions/33639/ssl-certificate-for-httpsosmorg-does-not-match-the-host-name/33640
[2]: https://wiki.openstreetmap.org/wiki/Domain_names

Change-Id: I40748ca3f3c306aba5010d433e8ae2c5369320a3