Use HTTPS for openstreetmap.org links
openstreetmap.org does not yet redirect to HTTPS by default [1], but
they seem committed to support HTTPS for the time being, so we can link
to their domain using HTTPS and trust that the link should not break in
the near future.

For the tile servers, this also enables the use of HTTP/2 [2], which
reduces load on the OSM servers and speeds up downloads for the user.

[1]: https://github.com/openstreetmap/operations/issues/117
[2]: https://github.com/openstreetmap/operations/issues/117#issuecomment-258642954

Change-Id: I7513a7edd28a9f3c5f515c69634f1707cdb7e67e