Add helper functions related to options
AbstractResultBrowser.getOption() returns a single option, falling back
to a provided default value if the option is not specified.

FormatterHelper.getColumnNamesOption() turns one or more variable names
into the corresponding column names. The variable names can be specified
as a single value or as an array, and should start with a question mark.
Example:

    #defaultView:Map{"layer": "?partyLabel"}

Change-Id: I78cfb4d8d53806146e3c27cef8a6aec7e35f9291