Use leaflet.markercluster for single-layer maps
Unless disabled in the options, use leaflet.markercluster for map views
with a single layer, which clusters nearby coordinates. The clusters
split up as you zoom in; coordinates that are still too near to each
other at the closest zoom level “spiderfy” (that is, they split up when
you click the cluster).

By default, for best compatibility with existing queries, we specify a
maximum cluster radius of 0 (spiderfy identical coordinates but don’t
cluster nearby ones) and create icons that closely resemble the ordinary
circle markers – the difference should only become apparent when you
click on a cluster.

It’s also possible to explicitly use the default options of
leaflet.markercluster, to cluster nearby markers and use a more
appropriate icon for the clusters:

    #defaultView:Map{"markercluster": {}}

The options can also be further changed, for example:

    #defaultView:Map{"markercluster": {"maxClusterRadius": 10}}

Currently, clustering doesn’t work well with map layers, so we disable
it by default in that case. If you want to see what happens, you can
explicitly enable it:

    #defaultView:Map{"markercluster": true}

Bug: T150136
Change-Id: Ib4ea273ab10c4b0bef95cb2035166dd47c7b8e65