Only initialize code example dialog when shown
We only need to initialize the code example dialog when the modal is
shown, not every time it is focused. In most browsers, the focus event
was only fired once when the modal was shown anyways, but in
WebKit-based browsers (e. g.  Safari, GNOME Web), the event seems to
behave slightly differently, which resulted in the dialog being
reinitialized with every click in its navigation area.

Bug: T179535
Change-Id: Ib47e91c620969bc2bdb425f3e11792db830e3d2b