Respect Do Not Track setting
If the user does not want to be tracked, don’t send tracking events.
Based on similar code in the WikimediaEvents extension [1], but supports
more browsers (the doNotTrack attribute is sometimes found on the
navigator and sometimes on the window, see MDN [2]) and only matches “1”
at the beginning of the string, since the header may contain extensions
(see specification [3]).

Also fixes a FIXME so that the method always returns a promise (even
though no caller currently uses the returned value).

[1]: https://github.com/wikimedia/mediawiki-extensions-WikimediaEvents/blob/1e68571e026db0f99e2c816f81224b83e89a2407/modules/all/ext.wikimediaEvents.statsd.js#L25-L29
[2]: https://developer.mozilla.org/en-US/docs/Web/API/Navigator/doNotTrack
[3]: https://www.w3.org/2011/tracking-protection/drafts/tracking-dnt.html#dnt-extensions

Change-Id: I01cdd4b5a6eeeb0ac70c33a630271a6ef960956d