Use moment’s relative time for “x seconds ago” message
Formatting the duration stand-alone and then appending some string does
not always yield correct results. For example, in German, the
stand-alone duration is in the nominative case („eine Minute“), but the
suffix requires the dative case („vor einer Minute“, not „vor eine
Minute“). Moment.js can fix this for us; unfortunately, it requires an
incompatible change to the message (Moment.js already adds the “ago”, so
we can’t have it in the message), so the message key is changed and we
lose all existing translations.

Change-Id: I08fd2133a1cd772e844fa4b73c5a9673662f6945