Add lazy loading and image rows
Lazily load images for the image result bowser. Display images in rows
of varying height, with every element in a row of the same height. This
prevents staggering at the bottom and makes the order of images clear.
Images load in parallel and rows display as soon as every image in the
row loads.

Bug: T166216
Change-Id: I8447b28c0db0015a257879dc77c1a74c44f3c62c