Support multiple languages in query templates
The "template" member of a query template can now not only be a string,
but also a map from language codes to template strings for that
language. When building the template, we search for the best template
string for the current language, respecting $.i18n language fallbacks.

The #TEMPLATE query hint is converted to object form, but without a
second translation. Hopefully the object form alone is enough of a hint
that you can add other language codes.

Bug: T188110
Change-Id: I27121b825b8e31e59d003f7064d1d7da4454423c