Only hide “scroll to top” in embed.html
The fix for T178222 in commit 07aa3535ba (change I53b82b1b5f) hides all
Leaflet controls in the top right corner, which is too heavy-handed – we
want to hide the “scroll to top” button, but not the layers control.
Instead, add a custom class to the “scroll to top” button and only hide
that button specifically, leaving the layers control in place.

Bug: T193930
Change-Id: I78bb522868a5d8b9eaf30148e8580f85413e8dcb