Extract _i18n helper functions into separate file
Instead of keeping half a dozen slightly different copies of an _i18n
helper function around, extract them into one function in a new file and
import that everywhere.

Also remove the use of I18N_PREFIX, since it makes it harder to
understand the code or find places where a certain message is produced.

Change-Id: I55ce0d514654b80a411de17893a77b43d7544bab