Fixes problem with WDQS HTML table export for cells without value
When downloading a result set in WDQS as HTML table, the cells that don't contain any values are not created as table cells in the HTML table by using <td> elements. This causes the cells to not properly be aligned when columns containing optional values are on the left or in the middle.

Bug: T207257
Change-Id: If6abb126b469dfcf53112b453628a46ffc714d8f