Try running code samples tests anywhere
Instead of preemptively deciding where running the code samples tests
will and won’t work, just try loading the code samples, and then
distinguish between different cases in the event of an error: when using
the file:// protocol, an error is to be expected and we let the test
pass, otherwise make it fail. Since different requests trigger the
failure, depending on browser, this logic is extracted into a function.

The effect of this is that the test suite should now pass regardless of
whether you use `grunt qunit`, open file://…/wikibase/tests/index.html
in a browser, or serve the whole project via `npm start` and then open
http://127.0.0.1:8080/wikibase/tests/index.html in a browser; however,
only in the last case will the code samples tests actually run. This is
not great, but it’s at least better than nothing.

Bug: T210503
Change-Id: I450916667eef9158237475826bfafd7314737c04