Do not sanitize a few more tooltips
Since version 3.4.1, Bootstrap sanitizes the content of tooltips [1]
(see also the previous commit message for more information). Sometimes,
this is nice: for example, the “data updated 2 minutes ago” popover was
previously vulnerable to an XSS attack by placing a <script> tag in a
message translation, which the Bootstrap update fixed. At other times
however, we properly control the HTML and would like Bootstrap to leave
it alone, and we signal this with a 'sanitize' option.

[1]: https://blog.getbootstrap.com/2019/02/13/bootstrap-4-3-1-and-3-4-1/

Change-Id: Id03677250d4fdb6fc66e345177f657cec8b9f4b4