Retry queries with POST if server indicates request was too large
Previously, we would only retry a query using POST if the browser
completely refused to send a query as a GET request. However, there’s a
window where the browser is still willing to send the request but the
server refuses to process it; in that case, we should still fall back to
a POST request.

Based on code provided by NoInkling in T216590#4967664.

Bug: T213889
Bug: T216590
Change-Id: I98958818655e3fe1df153677064924b2921414bc