Use an <iframe> for rendering of UrlShortnener._getWikiShort()
This is part of a larger bug, because we should not use popover
together with a dropdown menu. When clicking an item in a dropdown
menu, the menu wants to close.

But, it seems by luck, the "tinyurl" approaach with an iframe
works becase iframes have their own focus, thus, even when clicking
in the iframe (to select the text and copy it), the menu does not
see that the click and thus think nothing happened and stay open.

For consistency and as short-term fix, simply use the same
approach from TinyURL and apply it to MW UrlShortener as well.

The styling is also made to match (using <pre>).
Follows-up f5ed1c3dc1b.

Test Plan:
* On localhost, verify TinyURL works and is selectable via menu.
* Edit default-config.json, changing 'urlShortener' from 'tinyurl'
  to 'wikibeta.
* Refresh, and try again. It will fail to get a url (because
  we do not allow shortening of 'localhost' urls), but the error
  message will be selectable which proves that the patch works.

Bug: T221127
Change-Id: Ib06098c476c51c957cb18d1c175dda29fda23f2b