Fix two HTML injections
The message for “x results in y ms” used html() instead of text() to
process the &nbsp; entity, which we can do with a hard-coded replace
instead (no existing translation uses any other entity, nor do I expect
any future translation to start doing so).

The tooltip for entities combined the entity label and description
(unescaped) with some HTML code for the tooltip contents and then showed
the full result as HTML; instead, create a jQuery collection
immediately (safely escaping the text parts) and eventually use that
when creating the tooltip.

Bug: T238822
Bug: T238824
Change-Id: I2ceda6911d43464dba330d41c871a8d8617451a2