Add default user agent to some more code examples
With this, most of the programming languages for which we provide code
examples (as opposed to wiki link, Listeria, etc., where this doesn’t
apply) specify a default user agent and advise the user to adjust it.

JavaScript (jQuery or modern) doesn’t specify a user agent because
query.wikidata.org does not allow custom User-Agent headers for CORS
requests (its Access-Control-Allow-Headers preflight response does not
name User-Agent), so the browser’s default user agent must be used.

Bug: T226709
Change-Id: I1363acd78805c0f56f2126bfecefbfe02b67bfa7