Use CORS instead of jsonp for cross domain requests
CORS is generally better than jsonp in everyway. This is also
important to help implement CSP, as we want to limit the places
we can load scripts from.

For the statsv beacon case, this could cause a warning to show
up in debug console. The warning is fine, as we don't need to
read the result, and the request goes through fine. However the
warning is kind of ugly, we could add a cors header to the statsv
endpoint to get rid of it. Alternatively we could load the beacon
as an image or something instead of using AJAX.

Bug: T238618
Change-Id: If3464bfce65c34507b50993c7c46fc18431c9ef3