Don’t autofocus editor when embedded
When the query service editor (that means index.html, not embed.html) is
embedded into a page, it grabs focus as soon as it loads, potentially
scrolling the surrounding page until the cursor is in view. This is rude
to the surrounding page and usually not intended.

(One hopes that the surrounding page would eventually be able to protect
itself against this using a suitable feature policy for the iframe – via
the Feature-Policy header or the allow attribute on the iframe itself –
but that is still under discussion at the W3C [1].)

[1]: https://github.com/w3c/webappsec-feature-policy/issues/273

Bug: T245637
Change-Id: Iaf3f2a1542801164db340618eaea6130055564b6